requirements.txt
streamlit
pandas
numpy
plotly
statsmodels
matplotlib
seaborn
python-dateutil
